Create PressPackageItem component
