+++ components/admin/PackagePreview/PackagePreview.scss
@@ -47,67 +47,6 @@
         margin: 0;
       }
     }
-
-    .ui.card {
-      border-radius: 0.111111111rem;
-      @supports (display: grid) {
-        margin: 0;
-      }
-
-      .use-container {
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-        border-radius: 0.111111111rem 0.111111111rem 0 0 !important;
-        padding: 1em;
-        color: #ffffff;
-        background-color: $blue;
-      }
-
-      .icon {
-        height: 30px;
-        width: 30px;
-      }
-
-      .meta {
-        margin-top: initial;
-
-        > .date-time {
-          font-size: 12px; // match size of tags
-
-          > div > dt {
-            margin-right: 0;
-          }
-        }
-
-        > .modal_section--postTags {
-          margin-top: -0.75em;
-          margin-bottom: 1em;
-        }
-
-        > .seal {
-          color: initial;
-          font-family: Georgia, serif;
-        }
-      }
-
-      > .header > .title {
-        padding: 1em;
-        font-size: 0.777777778rem
-      }
-    }
-
-    .ui.cards > .card > .content,
-    .ui.card > .content {
-      border-top: 0;
-    }
-
-    .ui.cards > .card .meta,
-    .ui.card .meta {
-      padding: 1em;
-      text-align: left;
-      background: transparent;
-    }
   }
 }
